+++ src/main/kotlin/io/github/oxisto/reticulated/ast/expression/primary/atom/literal/StringLiteral.kt
@@ -15,7 +15,7 @@
  *
  */
 
-package io.github.oxisto.reticulated.ast.expression.literal
+package io.github.oxisto.reticulated.ast.expression.primary.atom.literal
 
 /**
  * This class represents a stringliteral
