+++ src/main/kotlin/io/github/oxisto/reticulated/ast/simple/SimpleStatementVisitor.kt
@@ -20,9 +20,9 @@ package io.github.oxisto.reticulated.ast.simple
 import io.github.oxisto.reticulated.ast.EmptyContextException
 import io.github.oxisto.reticulated.ast.Scope
 import io.github.oxisto.reticulated.ast.expression.ExpressionVisitor
-import io.github.oxisto.reticulated.ast.expression.AtomVisitor
-import io.github.oxisto.reticulated.ast.expression.Identifier
-import io.github.oxisto.reticulated.ast.expression.Name
+import io.github.oxisto.reticulated.ast.expression.primary.atom.AtomVisitor
+import io.github.oxisto.reticulated.ast.expression.primary.atom.Identifier
+import io.github.oxisto.reticulated.ast.expression.primary.atom.Name
 import io.github.oxisto.reticulated.ast.simple.target.TargetVisitor
 import io.github.oxisto.reticulated.grammar.Python3BaseVisitor
 import io.github.oxisto.reticulated.grammar.Python3Parser
@@ -52,7 +52,7 @@ class SimpleStatementVisitor(val scope: Scope) : Python3BaseVisitor<SimpleStatem
     val module = ctx
         .getChild(1)
         .accept(
-        AtomVisitor(this.scope)
+            AtomVisitor(this.scope)
     ) as Identifier
 
     // build a name
