+++ src/main/kotlin/io/github/oxisto/reticulated/ast/statement/FunctionDefinition.kt
@@ -19,7 +19,7 @@ package io.github.oxisto.reticulated.ast.statement
 
 import io.github.oxisto.reticulated.ast.Suite
 import io.github.oxisto.reticulated.ast.expression.Expression
-import io.github.oxisto.reticulated.ast.expression.Identifier
+import io.github.oxisto.reticulated.ast.expression.primary.atom.Identifier
 
 /**
  * A function definition
@@ -27,9 +27,9 @@ import io.github.oxisto.reticulated.ast.expression.Identifier
  * Reference: https://docs.python.org/3/reference/compound_stmts.html#grammar-token-funcdef
  */
 class FunctionDefinition(
-  val id: Identifier,
-  val parameterList: ParameterList = ParameterList(), suite: Suite,
-  val expression: Expression?
+    val id: Identifier,
+    val parameterList: ParameterList = ParameterList(), suite: Suite,
+    val expression: Expression?
 ) : Definition(suite) {
   // TODO: decorators
