+++ src/main/kotlin/io/github/oxisto/reticulated/ast/statement/Parameter.kt
@@ -19,7 +19,7 @@ package io.github.oxisto.reticulated.ast.statement
 
 import io.github.oxisto.reticulated.ast.Node
 import io.github.oxisto.reticulated.ast.expression.Expression
-import io.github.oxisto.reticulated.ast.expression.Identifier
+import io.github.oxisto.reticulated.ast.expression.primary.atom.Identifier
 import io.github.oxisto.reticulated.solver.Identifiable
 
 /**
