+++ src/main/kotlin/io/github/oxisto/reticulated/ast/statement/ParameterVisitor.kt
@@ -17,10 +17,10 @@
 
 package io.github.oxisto.reticulated.ast.statement
 
-import io.github.oxisto.reticulated.ast.expression.AtomVisitor
+import io.github.oxisto.reticulated.ast.expression.primary.atom.AtomVisitor
 import io.github.oxisto.reticulated.ast.Scope
 import io.github.oxisto.reticulated.ast.expression.ExpressionVisitor
-import io.github.oxisto.reticulated.ast.expression.Identifier
+import io.github.oxisto.reticulated.ast.expression.primary.atom.Identifier
 import io.github.oxisto.reticulated.grammar.Python3BaseVisitor
 import io.github.oxisto.reticulated.grammar.Python3Parser
 import org.antlr.v4.runtime.ParserRuleContext
